+++ CHANGES.rst
@@ -1,6 +1,9 @@
 Change log
 ==========
 
+**2023-07-11 (2.15.0)**
+* Add option to create municipality limits.
+
 **2023-06-29 (2.14.2)**
 * Upgrade certifi to fix catastro SSLCertVerificationError.
 * Upgrade to https urls (currently redirecting).